Create hook to check user state and component for generic page layout
