+++ frontend/src/pages/VulnerabilityScanDash/VulnerabilityScan.tsx
@@ -3,8 +3,8 @@ import { withSearch } from '@elastic/react-search-ui';
 import { Box, BoxProps, Grid, CircularProgress, Link } from '@mui/material';
 // Context & hooks
 import { ContextType, useAuthContext } from 'context';
-import useFirstLoginPopup from 'hooks/useFirstLoginPopup';
 import { useVulnScanData } from 'hooks/useVulnScanData';
+import { useCheckUserState } from 'hooks/useCheckUserState';
 // Constants & types
 import { ORGANIZATION_FILTER_KEY } from 'components/FilterDrawer/VSDashRegionAndOrgFilters';
 import { OrganizationShallow } from 'components/FilterDrawer/RegionAndOrganizationFilters';
@@ -13,8 +13,8 @@ import { LoginBlockedDialog } from 'components/LoginBlockedDialog';
 import { UpdateStateForm } from 'components/Register';
 import InfoLabel from 'components/Dashboard/InfoLabel';
 import NoDataMessage from 'components/Dashboard/NoDataMessage';
-import FirstLoginPopup from 'components/Dialog/FirstLoginPopup';
 import InvitePendingCard from 'components/Dialog/InvitePendingCard';
+import PageSection from 'components/Dashboard/PageSection';
 // Local dashboard components
 import KeyMetrics from './KeyMetrics';
 import ScanningSummary from './ScanningSummary';
@@ -56,39 +56,6 @@ const keyMetricsStyle = {
   alignItems: 'center'
 };
 
-const PageSection = ({ children }: { children: React.ReactNode }) => {
-  const { user } = useAuthContext();
-  const { show: showFirstLoginPopup, close: handleCloseFirstLoginPopup } =
-    useFirstLoginPopup(user ?? null);
-  return (
-    <Box
-      sx={{
-        maxWidth: '1152px',
-        margin: 'auto',
-        px: { xs: 0, sm: 0.5, md: 1, lg: 1, xl: 0 },
-        pb: 6,
-        minHeight: '100vh'
-      }}
-    >
-      <Grid container mb={0}>
-        <Grid size={{ xs: 12 }} sx={{ mt: 4 }}>
-          <InfoLabel
-            label="Vulnerability Scanning Dashboard"
-            typographyVariant="h1"
-            viewDetails
-            tooltipContentJson={tooltipContentJson}
-          />
-        </Grid>
-      </Grid>
-      <FirstLoginPopup
-        showFirstLoginPopup={showFirstLoginPopup}
-        handleCloseFirstLoginPopup={handleCloseFirstLoginPopup}
-      />
-      {children}
-    </Box>
-  );
-};
-
 export const VulnerabilityScan: React.FC<ContextType> = (props) => {
   const { filters, removeFilter } = props;
   const {
@@ -100,23 +67,14 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
     isLoggingOut
   } = useAuthContext();
 
-  const [isUpdateStateFormOpen, setIsUpdateStateFormOpen] = useState(false);
-
+  const { isUpdateStateFormOpen, setIsUpdateStateFormOpen } = useCheckUserState(
+    user,
+    isLoggingOut
+  );
   const [isLoginBlockedDialogOpen, setIsLoginBlockedDialogOpen] =
     useState(false);
   const [maintenanceNotification, setMaintenanceNotification] =
     useState<any>(null);
-
-  useEffect(() => {
-    if (!isLoggingOut && user) {
-      if (
-        (!user.state || user.state === '') &&
-        !localStorage.getItem('user_state')
-      ) {
-        setIsUpdateStateFormOpen(true);
-      }
-    }
-  }, [user, isLoggingOut]);
 
   useEffect(() => {
     const fetchAndCheckMaintenance = async () => {
